langgraph-solution/pyproject.toml
[project]
name = "langgraph-solution"
version = "0.1.0"
description = "Multi-agent LangGraph application for German energy market processes"
readme = "README.md"
requires-python = ">=3.14.2"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
energy-agents = "src.main:main"
